refactor - separate task actions and req thunks
