+++ src/redux/actions/habit-local-storage.svc.ts
@@ -1,0 +1,2 @@
+const LOCAL_STORAGE_PREFIX = 'MOMENTUM_HABIT_DATA_';
+
